+++ cld/entities.yml
@@ -15,6 +15,8 @@ AppContext:
         inputMDFiles: [Int: [String]]
         listInputDirId: Int
         projectPath: String
+        readFile: String
+        readFileContents: String
         request: NetRequest
         response: NetResponse
         responseError: NetResponse
